Create Testpage.html
I have been unable to reconcile the directory structure of my repository with the directory structure of the website Valcarcel.github.io. The directories in the repository are not accessible as webpages through a browser. This is a test at creating a new html file that can be viewed via a web browser.